修复executionGraphCache oom问题
